Fixes migration transaction lock isolation issue
Because the migration lock is obtained inside a transaction, it's possible for
the view of the database to be out of date once the lock is obtain (if another
process was migrating the database at the same time, but had not commited at
the time the lock was obtained). That could cause the next migrator to try
re-apply the same migrations, only to end up getting an error.

Now we *try* to obtain the lock. If we cannot, then we wait for the lock to
become open, but exit and obtain a *new* transaction with the current view of
the database. Of course, in that new transaction we have to apply the same
process again, until we end up actually obtaining the lock for real.