Allows use of abitrary haskell types as primary keys
This requires adding a type parameter to TableDefinition to capture the
primary key type, as well as conversion functions related to it.

Note that this is probably not actually all the useful with Poppers
unless a similar change is made to FieldDefinition to track the Haskell
type associated with the fields. I'll add that as another commit here.