Adds a fieldValue type parameter to FieldDefinition
This type represents the Haskell type that should be used for values
stored in the corresponding database column. This requires adding
another field to the FieldDefinition tuple that defines how to convert
the value to/from sql.

The bad news is that this breaks *every* FieldDefinition in every
project using Orville. The good news is that tugging on this thread has
eliminated about 80% of the usages of Convertible *and* those that
remain are mostly internal usages in Orville. The more good news is that
now Where constructs and Popper functions that used to look like:

    (.==) ::
          Convertible a SqlValue
       => FieldDefinition
       -> a
       -> WhereCondition

    hasMany ::
          TableDefinition entity key
       -> FieldDefinition
       -> Popper Record [entity key]

Now have better type safety built into them:

    (.==) ::
          FieldDefinition a
       -> a
       -> WhereCondition

    hasMany ::
          Ord fieldValue
       => TableDefinition entity key
       -> FieldDefinition fieldValue
       -> Popper fieldValue [entity key]

So now users are forced to give a Haskell value matching the type
specified by the FieldDefinition that is being used. That feels like a
big win.