Changes types to use app assigned primary key values
This essentially only changes the types involved in things so that
Orville doesn't apply the `entity` type variable to other variables.
Instead `entity` should be a fully applied concrete type.

Overall this makes the type signatures of functions simpler to read.
Occassionaly we have lost some clarity on how the `entity` and `key`
types are related. Only the presence of a `TableDefinition` in the
function signature enforces this relationship now.

We also lose any sense in the type system of "saved entities" vs
"unsaved entities". Where the type of `insertRecord` previously gave a
hint that entity given should not yet be saved now there is no such
hint.

This commit does not provide any facilities to help applications
generate approriate primary key values. More experimentation surrounding
that to follow.