Extends full/partial entity switch to RelationalMap
This introduces a minor change in the way maps for TableParams with
automatic ids are built. Now these need to be specifically listed as
read only portions of the map. That is a minor inconvenience, but seems
like a good thing for expressing clearly the flow of data in and out of
the database.

This also exposed a bug where fields marked as read only were not being
included in the list of fields for the table, and so were not being
added at all.