Adds support for application managed keys
This exploits the new flexibility of having separate type parameters for
read and update by specifying the *same* type for both, allowing
application to assign primary keys.

This required reworking a bit of the insert/update logic around columns
which previously had logic specific to primary keys. Now that logic
requires fields controlled by the database to be explicitly marked as
such. When using `RelationalMap`, this marking is done automatically by
`readOnlyMap`.