[ch12596] Adds Database.Orville.Trigger
This provides OrvilleTriggerT which can be used instead of OrvilleT.
With OrvilleTriggerT you can use the insertTriggered, updateTriggered
and deleteTriggered operations to have OrvilleTriggerT automatically
keep track of an application specific list of trigger events that
occurred in while running a block. The list of trigger events can then
be used for later processing, such as scheduling background jobs to
do extra processing in response to them.