[ch12596] Moves the startTransactionSQL callback into OrvilleEnv
This simplifies the typeclass and opens the door to more sophisticated
callbacks being registered in the OrvilleEnv datatype rather than flooding
the typeclass with many methods.

It's possible this is the beginning of a "Driver"-like framework, but it's
unclear. Mostly I want to see if this is workable from an end-user standpoint.