[ch12596] Moves runningQuery callback into OrvilleEnv
test/TestDB.hs provides an instructive example of the impact of
this change to users of the tracing feature. On the one hand I
think the user code is more straightforward overall, with the
tracing concern now being concentrated in just a couple of
functions. On the other hand, the tracing concern has been
completely eliminated from the type signatures, so the type
system cannot be used to make guarantees about it.