[ch12596] Adds TransactionEvent broadcasting
This adds a callback to allow Orville to let interested parties know
when transactions start,commit, and rollback. The callbacks are
monomorphized to IO for simplicity and clarity. The presumption is
that if you are a sophisticated enough user to be interested in these
callbacks that you will understand the value of dealing with any
necessary unlifting yourself rather than having Orville tie you to a
specific unlifting implementation.

This does, however, eliminate the possibility of using a stateful
monad (Writer, State) if your goal is simply to record a list of the
transaction events that happen. Instead you have to use an IORef (or
similar) construct that you can access from the IO monad.