Spikes unification of ColumnType and SqlConversion
This presents a proposed SqlType that would replace both
ColumnType and SqlConversion. Combining these two types
creates a single type that knows both the database column type
needed to hold a Haskell type *and* the Haskell mapping
necessary to serialize and deserialize column values from the
database.

Notably, by removing the ColumnType enum it should now be
possible to allow third parties to add support for new
database column types. This would also allow us to separate
out postgresql specific types (e.g. text search vector) into
the a separate postgresql library eventually.

There are still several problems to be solved before this is
complete, including how to determine the appropriate database
type when constructing a foreign field reference. Previously
we were able to map AutomaticId to ForeignId, that can no
longer be done in a rigorous way. I think that is a good
problem to solve because it paves the way for supporting other
kinds of primary keys in the foreign key reference helpers.