Adds generateMigrationPlan
generateMigrationPlan allows users to build and view the plan that Orville will use to migrate
a database schema without the plan being executed. This can be used simply to preview changes
at the repl during development, or to verify that the database you are connected to matches
your Haskell data definitions without attempting to modify it (e.g. you are accessing a read-only
database).