Adds support for lts-11/12 (including conduit-1.3)
This addresses both semigroup related issues and breaking changes in the
conduit library. For the time being I have left the old conduit code in place
and adding conditional compilation to maintain compatibility with older lts
versions. If this becomes cumbersome to maintain, we may decide to drop support
for older lts versions before the 0.9 release, but if it turns out to be
trivial we might as well keep it.