Backport: Fixes infinite loop bug when popMany is used on pure
The old handing of popMany for the PopPure case allow an infinite
list of results to be produced even when the list of inputs was
finite. This could result in an infinite loop when that infinite list
is used as the input to further operations (either within the scope
of the popper, or after the popper has finished and returned an
infinite list).

(cherry picked from commit bac808d9f63ef7ad601b989b5c542fe95db99de4)