Adds support for MonadUnliftIO
Rather than making this explicitly the default in the library by
exporting the UnliftIO oriented functions from the Core module, I
opted instead to built it as a peer to MonadBaseControl. This does
not represent a change in thinking that we should direct new users to
use UnliftIO as a default approach, merely that we should convey that
opinion through means other that module organization in the library.
For instance, doing it through both comments in the API documentation
and directing users to UnliftIO rather than MonadBaseControl in any
tutorial docs we write in the future.