Adds test-all script
This script will run compile and run the tests against all the
stack*.yml files that are found in the root of the repo, reporting
whether each one passed or failed.

It is generally useful for doing a final check before pushing, rather
then as the regular edit/compile/test loop to get feedback during
development.