Updates to text fields for Oracle.
On text(read: CHAR(n)) columns, Oracle will right pad inserted values with spaces.
This manifested an issue when doing direct comparisons, such as WHERE clauses, that
values did not compare in the expected behavior.

This changes the textField behavior to pad values on the way into the database
and strip whitespace on the way out.

Other minor change is moving an escaping helper to another module for reuse (when we finally add it back).
And an update to the sample project to reflect column names being capitalized in Oracle.

Notes/Gotchas:
  - Orville does not have support for the Oracle types: VARCHAR2 or NVARCHAR2 so this has not been tested against those.
  - The behavior of the textField now means that data purposely ending with some number of spaces will not retain
    those spaces on the way out of the database, but potentially many more would have been added in the previous behavior.
  - Perhaps add "rawTextField" to allow users of Orville to handle the whitespace issue in application code, if needed?

Things still broken/needing attention:
  - The test suite needs to be completely reworked with instructions on using the available Oracle 12c docker image.
  - Several functions still do not work including:
    - `migrateSchema` (causes db thread to spin indefinitely)
    - `selectFirst` (uses incompatible sql)
    - `insertRecord` ("Function sequence error")