Updates Oracle insertRecord to not give back the data just inserted.
This is due to a difference in how Oracle handles RETURNING. Notably, they use a
RETURNING INTO clause which is the area that needs further investigation.

Also includes in the README notes about functionality that has been tested/known to work,
as well as things that are known to not work.

Lastly, impacted areas in the sample-project remain commented out with the hopes that they are
 brought back shortly.