Initial generalization of local test Oracle setup, more work needed to get to full  test suite and complete reproducibility.
Adds the Oracle docker image to the orvile-oracle docker-compose file.
Note that this is not safe to run in CI because it requires a dockerhub account
and to be logged in via `docker login`.

Also adds a script to create an orville user in the database because the
Oracle image gives us no hooks to specify a non-system user for creation like
the PostgreSQL image does.

Finally, slightly adjusts Dockerfile for apt dependencies that are needed,
instructions for getting the closed-source Oracle library bits are still missing.