Adds Alternative and MonadPlus to OrvilleTriggerT
This adds a deriving clause for `Alternative` and `MonadPlus` to
`OrvilleTriggerT` (like `OrvilleT` hass) so that stacks it's used in can
use its underlying instances without having to (presumably) run their
respective OrvilleTriggerT's to unwrap them far enough to do so. We need
it in particular for a Happstack Monad stack.