Fixes `maybeMapper` breaking on already nullable fields
This changes `nullableType` to be aware of whether it is wrapping itself
around a type that is already nullable. In that case it uses the inner
type's 'NULL' decode rather than returning a `Nothing`. Essentially this
preserves as much information as possible about the value at the time of
decoding.