Replaces the 'Nullability' class with 'Nullability' value
This moves the core of the 'Nullability' concept to a field stored on
the 'FieldDefinition' while still tracking it at the type level.
'Nullability' is a GADT with associated type-level tags to promote the
value level tracking to the type level.

This both restricts the possible values for the 'nullability' type
variable more directly (because 'Nullability' values must be constructed
by the data constructors) and eliminates the need for the typeclass at
all. Where the typeclass was previously used pattern matching on the
value of 'fieldNullability' now provides the equivalent check.