Adds support for composite primary keys
This gives the `tablePrimaryKey` field of `TableDefinition` a rich
`PrimaryKey` data type that can represent composite primary keys and
updates the requisite places to make this work.

The `PrimaryKey` flag is now removed and the migration logic uses the
`tablePrimaryKey` field to determine how to create the primary key.
(Primary key migration is not currently support, and will not be until
we get moved to libpq at which point we may re-evaluate it).

The API remains relatively unchanged in most parts. The streaming
pagination needed to be changed to take a field defintion rather use
`tablePrimaryKey` to determine the ordering of the pagination because
using a composite primary key as the key for pagination will not work
due to ambiguity of ordering correctly on the multiple fields. On the
bright side, this means you can now pagination on any field you choose
rather than being restricted to the primary key field.