Makes Expr-types newtypes around RawSql
This introduces a new type, `RawSql`, for managing building raw sql
expressions. It's a thin wrapper around the bytestring `Builder` type
right now, though it may need to be expanded in the future to handle
values being put in for placeholders in prepared statements.

I changed all the types in `Expr` to be simple wrappers around the new
`RawSql` type rather than being direct models of expression structure.
The expression structure is now reflected in the smart constructors that
build expr values rather than the values themselves. This makes it easy
for us to expose a constructor to inject user-defined `RawSql` at any
level of the expression tree.

I also re-wrote the decoding tests to not duplicate the test structure
to avoid having to update each test every small change to `Expr` things.