LibPQ: Adds SqlValue to represent values exchanged with DB
This adds a new `SqlValue` type to be used when dealing with values that
are passed to or received from the database. The various encoding
function that lived as helpers in the `SqlType` module have been
promoted to functions exported from `SqlValue`.

This also exposed that the handling of sql NULL values was not correct.
NULL values are reflected a null pointeds in `libpq` rather than actual
string values, so `SqlValue` has a dedicated `SqlNull` constructor to
handle them.