Make FromSqlError Structured
Rather than free text, error messages now have a descriptive structure
which then allows us to have different printers for it based on whether
we are printing for logging purposes or if the is a potential for
sensitive information to leak out through the `Show` instance.
The printer used for the Show instance should produce the same message
as it did before.