LibPQ: Adds multi-insert and where clauses to `Expr`
This adds a bit more support to the `Expr` module with types and
constructors for building multi-row insert statements and an initial cut
at basic where comparisons.

Along the way I added a new `readRows` function that reads the rows from
the `LibPQ` result into `SqlValue` lists for further consumption by the
caller. I imagine this will get modified eventually based on the needs
of `SqlMashaller`.