LibPQ: Hooks SqlMarshaller up to LibPQ Result Sets
This adds an `ExecutionResult` typeclass whose functions mimic those of
LibPQ's `Result` type. The `SqlMarshaller` functions were then changed
to take use types that that satisfy this typeclass, which required
changing the how it interacts with the source data to match LibPQ's
access pattern.

I've provided `FakeLibPQResult` implementation of `ExecutionResult` as
well so that we can continue to write tests that don't have to touch the
database, but can still closely mimic lower layers of the integration
that have already been tested.