LibPQ: Checks for NUL bytes before sending value to LibPQ
This adds a new type, `PGTextFormatValu,e` to track whether we know that
the bytes it contains are all non-NUL (e.g. because we constructed
directly). For any values we do _not_ know are non-NUL, we check them at
the time of building the param list to pass to lib and abort running
the query. For now `executeRow` return `Nothing`, the same as if LibPQ
returned an error. We have another story to start doing better error
handling in general, so I have left actually raising the error about
NUL bytes for that story.