LibPQ: Surfaces SQL Execution Errors
This raises errors that can happen during exection as `Exception` values
in the `IO` Monad. These can be errors detected on the client side (e.g.
connection closed, null bytes found in input parameters) or returned
from the database sql (e.g. sql syntax error).

Once we start adding the high-level interfaces (in particular, using
orville within monad stacks) we may decide we want this lower level
interface to offer non-IO based error handling. We can address that
question once we get to it.

Side note: this removes the `stack.yaml` symlink because `docker-compose
up` was failing with a complaint that it could not find the file.