Breaks apart Expr module
This tries to find a balance between logical groupings of code and premature abstraction.
As with everything else I highly suspect we'll revist this many times as knowledge and needs shift.
The heart of this is to make the code a bit easier to digest and thus modify.

Some more details of changes:
- Makes Expr a re-exporter module
- Adds an Expr/Query hierarchy with the relevant bits for creating query expressions
- Adds an Expr/InsertExpr module with the insertion expression bits
- Adds an Expr/OrderBy hierarchy with all the OrderBy express code
- Adds an Expr/Where heirarchy with all the Where expression code
- Adds an Expr/Name heirarchy containing the Column and Table naming code
- Adds an Expr/FieldDefinition module
- Updates the years in modules to say 2021

FieldDefinition currently _feels_ a bit out of place with the others.
But I suspect future work with tables will likely resolve this.