Formats the LibPQ codebase with fourmolu
Running the formatter is scripted and depends only on POSIX, `stack`(to install fourmolu), and the `-P` extension to xargs which seems reasonably available and is used to run formatting in parallel.
For example it can run done with `docker-compose run --rm dev sh -c ./scripts/format-repo.sh`.

Some more details:
- Adds a fourmolu configuration file.
- Adds helper script for running the fourmolu during development (that will format all the .hs files found)
- Adds a CI job to check that formatting was done, which can be expanded for more tooling later if desired
  - This also uses a helper script, both placed in a orville-postgresql-libpq/scripts directory
- Adds a docker-compose file for running tools in CI and caching the build of them as the build of fourmolu was otherwise taking >2x the runtime of the other jobs
- Runs the formatting on the entire LibPQ codebase.