LibPQ: Adds minimal port of TableDefinition
This adds a minimal port of `TableDefinition`, containing just the table
name, marshaller and primary key. It includes functions to generate SQL
expressions for table creation, record insertion and querying.

`SqlMarshaller` needed some API tuning to get everything that
`TableDefinition` needed for its job. The "marshall to sql" function
became simply a fold over the fields in the `SqlMarshaller`, which is
used in a couple of different ways for building the queries to support
table operations.

We may decide that we don't want to export the `TableDefinition`
constructor (so that we can add record fields without it being a
breaking api change), but I deferred that decision for the moment in
favor of getting this basic functionality working.