Un-haddockizes typeclass method comments for fourmolu
`fourmolu` seems to crash when a haddock comment is placed on a typeclass
method, regardless of the format.

The comment on the `HasOrvilleState` class itself was also a problem,
but once I got `fourmolu` to format the file once I was able to add the
haddock back to that one.