LibPQ: Adds full support for field nullability
This adds the rest of the field nullability support from the HDBC
version. The main different is that I chose to not to export the GADT
for nullability outside the module itself. I think this probably makes
for a cleaner and less confusing API overall. The `FieldNullability`
type is the equivalent of `NullabilityCheck` from the HDBC code,
allowing users of `FieldDefinition` to recover the type-level
nullability information without needing to turn on GADTs.