LibPQ: Adds a build.sh for local development
New build.sh script does the following:
  - ensures the docker image is up to date
  - runs the tests against all the stack resolvers in the repo
  - runs formatting on the whole repo

It can be run like `sh scripts/build.sh` but it does need to be
run from within the libpq directory so the docker-compose commands
work out correctly.

Further it includes an option to be slightly more verbose
as well as an option to skip building the docker image.
This serves to save a small amount of time, as well as a template
should a future step be expensive.

Using `sh scripts/build.sh --help` prints the options.
Which at this point are hopefully descriptive enough to not
require further description.

This also moves the `test-all` and `test-loop` to live under the
scripts directory, which are both still available for direct running.

Also removes a stale file that was brought over from the HDBC side,
but that has never been used on the libpq code.