Orville-0.9: Remove libpq code
This removes the orville-postgresql-libpq directory entirely as it is the 0.10 code.
It not intended to be part of the 0.9 stable branch.

Further this removes the circleci config as thsoe jobs are only setup for the libpq code and no need to waste ci time on jobs that will always fail.