Ports the Select module from HDBC Orville in prep for Plan
This ports the `Select` module for building safer select custom select
statements from HDBC orville. `Select` the is basis for the primitive
plan operations, so this needed to be ported.

The `Select` modure replaces the `mkQueryExpr` function that previously
lived in the `TableDefinition` module. I imagine we will add similar
`Insert`, `Update` and `Delete` modules to replace the other query
building fuctions from `TableDefinition` as well.

This fixes `findEntitiesBy` not passing all it's `SelectOptions` parts
through as well, by virtue of using the new `Select` module.

A couple of other minor things ended up in this as way that I decided
not to tease so for the sake of getting this in this morning:

* `marshallNested` was added
* `whereIn` and `whereNotIn` renamed to `fieldIn` and `fieldNotIn`