Add Plans to LibPQ Orville
This is mostly a straight copy of the plan code from HDBC. The
`Operation` module needed to be adjusted a bit to match the new API, and
a couple of spots required `NonEmpty` handling, but otherwise the code
is unchanged.

The tests are completely new, however, and I think better targeted at
the plan functions.