Changes the naming of timestampFields for clarity.
When reading a column with a timezone we use the UTCTime type. The
field definition is now called `utcTimestampField`. Without a timezone we use the
`LocalTime` type. The field definition is now called
`localTimestampField`.

Note that the definitions in `SqlType` and `ColumnDefinition` are
still named to closely match the type names that PostgreSQL uses.