Adds modules to query 'pg_catalog'
This adds modules for several 'pg_catalog' tables and a higher-level
'DatabaseDescription' for describing a number of tables at once using
'Plan'.

I had to augment 'Expr' an 'Plan' a bit to support new types, including
the ability to query using tuples. Although the 'Expr' module supports
any size tuple, it does not enforce that the caller using same-sized
tuples when building the queries.

In the higher order 'WhereCondition' and 'Operation' modules, I limited
the support to 2-tuples to easily enforce this without making the types
too crazy for the time being. We will want to iterate on this prior to
release.

For now, I stopped short of exposing any of this in 'Plan' directly.
Instead you currently have to reach down to 'Operation' to get the tuple
support.