Adds AutoMigration of column types
This uses the `pg_attribute` information from `pg_catalog` to determine
whether the type of a column needs to altered, or nullability added or
removed.

I also revised `SqlType` a bit to match the way information is actually
used during migrations, including _NOT_ needing the length of integer
types to know whether they need to be migrate (the OID is sufficient).