Adds OrderBy as a nicer layer over OrderByExpr for SelectOptions
This adds an `OrderBy` wrapper around `OrderByExpr`, similar to how
`WhereCondition` wraps `BooleanExpr`. Helper functions for building
`OrderBy` based on column names and field definitions are providing,
eliminating the need for the user to build the `OrderByExpr` directly.
Escape hatch functions are also provided for users that need more direct
access to the `Expr` layer.