Adds constraints to `DatabaseDescription`
This adds a `PgConstraint` type and related table definition to allow
querying the `pg_catalog.pg_constraint`. The `RelationDescription` type
in `DatabaseDescription` now includes a list of constraints in its
description. These constraints will be used in an subsequent commit to
support adding and dropping foreign keys during auto migration.