Update compiler versions tested against.
Adds ghc-8.10.6 and 8.10.7 as tested compiler versions.
Also updates our nightly check for ghc 9.0.1.

Something of note: a change in the way happy is built means that it
now requires itself to build. This will cause builds to fail for us as
the build tools do not solve for this. To work around it I've included
installing a version of happy from the package manager which allows
the transitive dep of happy to build.