Adds support for creating and dropping indexes
This adds `IndexDefinition`, similar to `ConstraintDefinition`, which
can be added to a `TableDefinition` to define unique and non-unique
indexes.

Any indexes found that are not included in the `TableDefinition` auto
automatically dropped in the same way constraints are dropped. Indexes
automatically created by PostgreSQL to support constraints not
considered for dropping.