LibPQ: Adds support for column default values
This adds the ability to set default values for `FieldDefinition`.
Default values set on fields will then be reflected in the database
schema as part of migrations.

There is a significant potential gotcha will creating custom default
values -- in order to determine whether migration is necessary, Orville
must use PostgreSQL's `pg_get_expr` function to decompile the expression
and compare it to the default specified on the `FieldDefinition`. The
expression returned by `pg_get_expr` for any given input can often be
surprising. We will want to follow this up with some better help to
ensure the result of migrations in a fixed-point for Orville's migration
logic and provide useful error output to help the user identify why when
it does not.

This required two new low-level features from Orville:

* String literal support for the `RawSql` type, which in turn required
  escaping. LibPQ provides escaping functions, but the require a
  `Connection` to use which complicates the code a bit.

* Support for marshalling a sql expression during selections rather than
  a regular column. `SyntheticField` was added to support this.