LibPQ: Adds error information to MarshallError
This adds detail error information to the `MarshallError` produced by
`SqlMarshaller` when marshalling fails. This includes row identifiers
for the row being decode, as well as details about the column and values
being decode and any more general message provided by the source of the
decoding failure.

A new `ErrorDetailLevel` type controls the level of detailed actually
included in when the `MarshallError` is rendered to a `String` to be
displayed. This can be specified manually simply by catching the error
and providing the desired detail level to the render function. A default
value can be provided as well when constructing the `OrvilleState`,
which is what will be used if the `MarshallError` is rendered as an
`Exception` (via its `Show` instance).

The `SqlMarshaller` lacks any notion of which columns might be the row
identifiers to inlude in the error message. A new wrapper type, the
`AnnotatedSqlMarshaller` now captures that information. This new type
is automatically created when the `TableDefinition` is built, using
the primary key of the table as the row id.